Refactor DB configuration
There's a little bit of a chicken-and-egg problem where to build the
docker image you need postgresql, but postgresql is provided by docker.

This leaves H2 for docker image building, but uses postgresql with
docker-compose and Travis still.  If we ever end up using some specific
Postgres features we can revisit this and maybe figure something else
out.